+++ httpbin/handlers.go
@@ -6,7 +6,6 @@ import (
 	"compress/zlib"
 	"encoding/json"
 	"fmt"
-	"math/rand"
 	"net/http"
 	"strconv"
 	"strings"
@@ -266,13 +265,13 @@ func (h *HTTPBin) Unstable(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	// failureRate
+	// failure_rate
 	var failureRate float64
-	rawFailureRate := r.URL.Query().Get("failureRate")
+	rawFailureRate := r.URL.Query().Get("failure_rate")
 	if rawFailureRate != "" {
 		failureRate, err = strconv.ParseFloat(rawFailureRate, 64)
 		if err != nil || failureRate < 0 || failureRate > 1 {
-			http.Error(w, "invalid failureRate", http.StatusBadRequest)
+			http.Error(w, "invalid failure_rate", http.StatusBadRequest)
 			return
 		}
 	} else {
@@ -1005,22 +1004,4 @@ func (h *HTTPBin) Bearer(w http.ResponseWriter, r *http.Request) {
 		Token:         tokenFields[1],
 	})
 	writeJSON(w, body, http.StatusOK)
-}
-
-// Returns a new rand.Rand from the given seed string.
-func parseSeed(rawSeed string) (*rand.Rand, error) {
-	var seed int64
-	if rawSeed != "" {
-		var err error
-		seed, err = strconv.ParseInt(rawSeed, 10, 64)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		seed = time.Now().Unix()
-	}
-
-	src := rand.NewSource(seed)
-	rng := rand.New(src)
-	return rng, nil
 }
