+++ static/index.html
@@ -104,7 +104,7 @@
 <li><a href="/status/418"><code>/status/:code</code></a> Returns given HTTP Status code.</li>
 <li><a href="/stream-bytes/1024"><code>/stream-bytes/:n</code></a> Streams <em>n</em> random bytes of binary data, accepts optional <em>seed</em> and <em>chunk_size</em> integer parameters.</li>
 <li><a href="/stream/20"><code>/stream/:n</code></a> Streams <em>min(n, 100)</em> lines.</li>
-<li><a href="/unstable"><code>/unstable</code></a> Fails half the time, accepts optional <em>failureRate</em> float and <em>seed</em> integer parameters.</li>
+<li><a href="/unstable"><code>/unstable</code></a> Fails half the time, accepts optional <em>failure_rate</em> float and <em>seed</em> integer parameters.</li>
 <li><a href="/user-agent"><code>/user-agent</code></a> Returns user-agent.</li>
 <li><a href="/uuid"><code>/uuid</code></a> Generates a <a href="https://en.wikipedia.org/wiki/Universally_unique_identifier">UUIDv4</a> value.</li>
 <li><a href="/xml"><code>/xml</code></a> Returns some XML</li>
